Add files via upload
Setting up the HTML
Adding JavaScript
First, we will get the current time using the Date object and store it in the now variable. Using the now variable, let's get hours, minutes, seconds, day, month, and year. Refer to the code below for a clear understanding.
Styling the digital clock